requirements.txt
gdown==4.5.3
lightly==1.2.27
matplotlib==3.5.3
numpy==1.21.6
omegaconf==2.3.0
pandas==1.3.5
Pillow==9.2.0
pytorch_lightning==1.6.5
scikit_learn==1.0.2
PyTDC==0.4.1
torchcontrib==0.0.2
torchmetrics==0.11.4
transformers==4.21.1
wilds==2.0.0
setuptools==58.0.4
yacs==0.1.8
einops==0.6.1
POT==0.8.2

